Refactoring and tests changes
