+++ trustpoint/request/request_validator/__init__.py
@@ -1,1 +1,8 @@
 """Initialization for the request validation step of the request pipeline."""
+
+from http_req import CmpHttpRequestValidator, EstHttpRequestValidator
+
+__all__ = [
+    'CmpHttpRequestValidator',
+    'EstHttpRequestValidator',
+]
